fixed bugs and refactored code

+++ src/main/java/org/app/service/BookService.java
@@ -36,15 +36,11 @@ public class BookService {
         try {
             for (Book book : bookRepo.retrieveAll()) {
                 Field bookField = book.getClass().getDeclaredField(bookFieldToRemove);
-
                 bookField.setAccessible(true);
                 Object bookValue = bookField.get(book);
-                //FIXME двойная вложенность if
-                if (Objects.nonNull(bookValue)) {
-                    if (bookValue.toString().equals(bookFieldValueDto.getInputBookFieldValue())) {
-                        bookRepo.removeItemByField(bookFieldToRemove, bookFieldValueDto.getInputBookFieldValue());
-                        logger.info("book is removed: " + book);
-                    }
+                if (Objects.nonNull(bookValue) && bookValue.toString().equals(bookFieldValueDto.getInputBookFieldValue())) {
+                    bookRepo.removeItemByField(bookFieldToRemove, bookFieldValueDto.getInputBookFieldValue());
+                    logger.info("book is removed: " + book);
                 }
             }
         } catch (NoSuchFieldException | IllegalAccessException ex) {
